+++ opentama/__init__.py
@@ -21,4 +21,4 @@ __all__ = [
     "save_state",
 ]
 
-__version__ = "0.3.2"
+__version__ = "0.3.3"